Prepare program for future subcommands
